Removing build-action
