+++ prospero-cli/src/main/java/org/wildfly/prospero/cli/CliMain.java
@@ -20,7 +20,6 @@ package org.wildfly.prospero.cli;
 import org.jboss.logging.Logger;
 import org.wildfly.prospero.actions.Console;
 import org.wildfly.prospero.cli.commands.ApplyUpdateCommand;
-import org.wildfly.prospero.cli.commands.BuildUpdateCommand;
 import org.wildfly.prospero.cli.commands.ChannelCommand;
 import org.wildfly.prospero.cli.commands.CliConstants;
 import org.wildfly.prospero.cli.commands.CloneCommand;
@@ -72,7 +71,6 @@ public class CliMain {
 
         commandLine.addSubcommand(new InstallCommand(console, actionFactory));
         commandLine.addSubcommand(new UpdateCommand(console, actionFactory));
-        commandLine.addSubcommand(new BuildUpdateCommand(console, actionFactory));
         commandLine.addSubcommand(new ApplyUpdateCommand(console, actionFactory));
         commandLine.addSubcommand(new HistoryCommand(console, actionFactory));
         commandLine.addSubcommand(new RevertCommand(console, actionFactory));
