+++ prospero-cli/src/main/resources/UsageMessages.properties
@@ -43,8 +43,10 @@ prospero.install.usage.customSynopsis.5 = \u0020        (to install from a galle
 prospero.update.usage.header  = Apply the latest available patches on a server instance.
 prospero.update.usage.customSynopsis.0 = @|bold ${prospero.dist.name} update|@ [@|fg(yellow) --dir|@=@|italic <path>|@] @|fg(yellow) [OPTION]|@...
 prospero.update.usage.customSynopsis.1 = \u0020        (to update a server installation)
-prospero.update.usage.customSynopsis.2 = \u0020 or:  @|bold ${prospero.dist.name} update|@ @|fg(yellow) --self|@ [@|fg(yellow) --dir|@=@|italic <path>|@] [@|fg(yellow) OPTION|@]...
-prospero.update.usage.customSynopsis.3 = \u0020        (to update ${prospero.dist.name})
+prospero.update.usage.customSynopsis.2 = \u0020 or:  @|bold ${prospero.dist.name} update|@ [@|fg(yellow) --dir|@=@|italic <path>|@] @|fg(yellow) --update-dir|@=@|italic <path>|@ @|fg(yellow) [OPTION]|@...
+prospero.update.usage.customSynopsis.3 = \u0020        (to prepare an update in a new folder)
+prospero.update.usage.customSynopsis.4 = \u0020 or:  @|bold ${prospero.dist.name} update|@ @|fg(yellow) --self|@ [@|fg(yellow) --dir|@=@|italic <path>|@] [@|fg(yellow) OPTION|@]...
+prospero.update.usage.customSynopsis.5 = \u0020        (to update ${prospero.dist.name})
 
 prospero.history.usage.header = List previous installation states.
 prospero.revert.usage.header  = Reverts to a previous installation state.
@@ -124,6 +126,7 @@ version = Print ${prospero.dist.name} version and exit.
 yes = Perform the operation without asking for a confirmation.
 path = Path of the file to export to or import from.
 update-dir = Path of previously prepared server update to be applied.
+prospero.update.update-dir = Path to generate the updated server into. The original server will not be changed.
 
 #
 # Exit Codes
